+++ src/django_orca/admin.py
@@ -1,8 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.models import Permission
 
-from .models import RolePermission, UserRole
+from .models import UserRole
 
 admin.site.register(Permission)
-admin.site.register(RolePermission)
 admin.site.register(UserRole)